+++ internal/tickets/dtos/create_ticket.go
@@ -7,7 +7,7 @@ type CreateTicketRequest struct {
 	NamespaceID uuid.UUID   `json:"namespace_id" binding:"required,uuid"`
 	QuotaID     uuid.UUID   `json:"quota_id" binding:"required,uuid"`
 	Resources   []Resources `json:"resources" binding:"required"`
-	Duration    float32     `json:"duration" binding:"required,gte=1"`
+	Duration    uint        `json:"duration" binding:"required,gte=1"` // in seconds
 }
 
 type Resources struct {
